Remove DCP operations from the pipeline when they are completed
We need to make sure we remove these operations otherwise we can
end up with the DCP handler getting called multiple times. This
will likely also cause memory issues if too many of these pile up.
Since getting a DCP stream end is a special case the best way to
handle this is to just cancel the operation once the we recieve
a stream end.

Change-Id: I34a65c62ab2272493ab93792bfc5e4c319023ddd
Reviewed-on: http://review.couchbase.org/56903
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>