Use the pool/default/bs as the streaming API for buckets
We need to use the bs streaming API instead of bucketsStreaming in
order to get the correct information for buckets. This fixes an issue
where view and n1ql query were unable to find nodes in the cluster
even though they exist. I also removed the fields that won't come
back in the streaming API response in order to make sure that they
aren't used on accident later.

Change-Id: Ic8d389e0026d25c3e5ef15bf398fb60f98d71fe0
Reviewed-on: http://review.couchbase.org/57444
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>