Add CouchbaseMock functionality for tests
This adds the ability to use the CouchbaseMock mock server for automated
testing. The server is downloaded (if not available locally) and then
executed.

The `agent_test` tests have now been re-enabled as they should be able
to run using the Mock and no longer require an actual Couchbase server.

Change-Id: If71e33a151224d4d963031a026e9181d4adb1baf
Reviewed-on: http://review.couchbase.org/58803
Tested-by: Mark Nunberg <mark.nunberg@couchbase.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>