Ensure that memdQRequests have nil pointers for queueNext
If we requeue an operation it's possible that queueNext already
has a value set for it. As a result we should make sure to set
queueNext to nil since we always insert at the end of the linked
list. This resolves a deadlock issue where the last item in the
memdqueue was pointing to a previous item in the queue. If this
happens it can cause the drain function to loop infinitely and
deadlock the client on shutdown.

Change-Id: I943a56a49169cfa46f007e225553330c2db23ff7
Reviewed-on: http://review.couchbase.org/61677
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>