GOCBC-210: Added support for subdoc full document delete.
Motivation
----------
The server has implemented the ability to perform full document
deletes within a set of subdoc operations.  While this functionality
is not generally useful, the Sync Gateway team has a use-case for
this.

Changes
-------
Added interception of blank paths in subdoc Remove method which
rewrites the operation into a fulldoc remove instead.

Change-Id: I903413f2b1397be94e7d25f4d6c6990365c61204
Reviewed-on: http://review.couchbase.org/80076
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Adam Fraser <adam.fraser@couchbase.com>
Tested-by: Brett Lawson <brett19@gmail.com>