GOCBC-162: Implemented KV error map.
Motivation
----------
The server now supports sending an error map which provides
retry information about errors as well as textual error information
that the clients can provide the user in the case of errors which
the client is not yet aware of (new server versions introducing
new errors).

Changes
-------
Updated the client to enable KV error map usage in gocbcore.

Change-Id: I12a1490eed83b1312e575567eb64995d6f8680d8
Reviewed-on: http://review.couchbase.org/80964
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Brett Lawson <brett19@gmail.com>