GOCBC-215: Refactor errors objects to support extra information.
Motivation
----------
The current kv error objects simply store the error code of the
error that occured.  This does not allow for dynamic error messages
or additional contextual information that may be available about
the error.  Since dynamic error information will be implemented
soon, the errors need to be refactored to allow additional data
to be stored.

Changes
-------
Updated internal error handling code to take advantage of new
status code based errors.  This is primarily to make it simpler
to backport v2 changes on to v1.

Change-Id: Idcf75be60bf5602111396261d1a4ef12789b90c5
Reviewed-on: http://review.couchbase.org/80965
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>