GOCBC-245: Added support for Ping and HealthCheck.
Motivation
----------
To aid developers and operations to more accurately determine
the state of the client and perform monitoring of their
applications, we are implementing the ability to fetch the
state of the clients through HealthCheck and Ping methods.

Changes
-------
Implemented HealthCheck and Ping functions which allow both
transient and active fetching of client state.

Change-Id: Id97914e3f8a983302b4860a08385bec0dc332c2b
Reviewed-on: http://review.couchbase.org/87077
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>