GOCBC-257: Implement log redaction.
Motivation
----------
Due to ongoing changes in the space of user information security,
in many cases, organizations are in need of the ability to ensure
that various pieces of data from their infrastructure are not
written to log files.  We've implemented the ability to automatically
redact this information from logs.

Changes
-------
Added a SetRedactionLevel method to the client which enables you to
specify the degree you wish to redact that logs.

Change-Id: I41cc46a4ab6e619470ccfad9185b7e933c2bd147
Reviewed-on: http://review.couchbase.org/90164
Reviewed-by: Ellis Breen <ellis.breen@couchbase.com>
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>