GOCBC-265: Added built-in threshold tracer.
Motivation
----------
In the majority of simple cases, users wish to be immediaately
aware of performance degradations within their application and/or
the cluster.  The threshold log tracer automatically prints
information to the log when these operations are detected.

Changes
-------
Implemented a new tracer which automatically generates log
messages on a recurring basis which includes information about
the slowest operations that occurred above a set of predefined
operation thresholds.

Change-Id: I435ac92f4a1f688c025bb0aa436024e6cec4ca2b
Reviewed-on: http://review.couchbase.org/90166
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>