GOCBC-278: Added profile option for N1QL queries.
Motivation
----------
Couchbase Server 5.0's query added a profile HTTP parameter
that takes a few different strings. To ensure this is
discoverable by developers, adding it to the APIs as a
parameter.

Changes
-------
Added a new method to the N1QL query builder which allows
you to specify what profile mode to use.

Change-Id: Ia8e501ed956a80690a1c8d6d6df19118f51e3fdb
Reviewed-on: http://review.couchbase.org/91462
Reviewed-by: Matt Ingenthron <ingenthr@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>