GOCBC-289: Correctedly propagate CAS for replace ops.
Motivation
----------
A bug was discovered where the CAS value being passed to
replace was not being propagated to gocbcore and thus not
being used.  This caused Replace operations to
unconditionally replace a document, even if a CAS check
was expected.

Changes
-------
Added correct propagation of CAS to gocbcore in Replace
method.

Change-Id: I78ba774044201cff977d31ba12e763bb42ad26a0
Reviewed-on: http://review.couchbase.org/92361
Reviewed-by: Matt Ingenthron <ingenthr@gmail.com>
Tested-by: Brett Lawson <brett19@gmail.com>