GOCBC-302: Add a check and error for mixed authentication
Motivation
----------
The use of a cluster-level authenticator alongside a bucket
password does not make a lot of sense.  We should be returning
an error in this case to force the user to clarify which
is intended to be used.

Changes
-------
We now check if both an authenticator and bucket password
were specified when opening a bucket and return an error if
this is the case.

Change-Id: I8d61abfdc0ebe15733a99bcde6c960fac14ca350
Reviewed-on: http://review.couchbase.org/94416
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>