GOCBC-230: Add support for mock and real server integration testing
Motivation
----------
The test suite should allow execution against the CouchbaseMock server
or a real server if specified.

Changes
-------
Added flags to bucket_test to allow for the details of a real
server to be passed from the command line. Also added
a function to allow tests to check if the server is a mock
instance. This function should be superceded by version checking.

Change-Id: I433fc65c34998cca3bf67af85423f5b004b1e1ab
Reviewed-on: http://review.couchbase.org/94693
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>