GOCBC-301: Fail fast when cert auth is setup but authenticator not used.
Motivation
----------
When gocb is initialized with cacertpath, certpath and keypath it's
"obvious" that client certificate authentication is going to be used.
We should check if the Authenticator being used if a CertAuthenticator
and fail fast if not.

Changes
-------
Added a new error and a check to ensure that CertAuthenticator is being
used if there are certificates present.

Change-Id: I7e034e0ad5a67d337f841b98ed879d17462d0990
Reviewed-on: http://review.couchbase.org/94756
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>