GOCBC-229: Add support for cluster-version-variable integration testing.
Motivation
----------
The SDK should implement the ability to pass the testing server version
to its test invocation. This will allow the SDK to perform only the
tests that are available to a particular server version and avoid tests
for features which are not yet implemented.

Changes
-------
Added a flag for passing server/mock version on the cli during testing.
Added functionality for checking if features are supported
by the test instance (be it real or mock)

Change-Id: Id64594e51a56c88a5f976cd715a79dd99fdc942e
Reviewed-on: http://review.couchbase.org/94772
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>