GOCBC-317 Don't pass CAS value to TouchEx
Motivation
----------
Touch should not send a non-zero CAS value to the server. As of
server 5.x this is enforced by the server and will return an error.

Changes
-------
Don't pass CAS from touch to TouchEx and change signature of touch,
leaving signatute of Touch and TouchDura as is. Fail fast in Touch
and TouchDura if a non-0 CAS value is provided.

Change-Id: I4f763ad276482346cf9955e47b9aa2e25b614b63
Reviewed-on: http://review.couchbase.org/95807
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>