GOCBC-341: Fix ExecuteSearchQuery json unmarshal error
Motivation
----------
When a malformed FTS query is sent to the server then the response
is a different format to a well formed query. When this happens
the SDK is not able to correctly parse the response.

Changes
-------
Add handling for when the FTS response status code is 400 or 401.
In these circumstances add the error to the SearchResults errors
to bring in line with the RFC.

Change-Id: Ic31d0cdf477a1efef8831e5832880ebc70b94c78
Reviewed-on: http://review.couchbase.org/98300
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>