GOCBC-347: Retry on HTTP 429 for FTS
Motivation
----------
As part of recent server changes in FTS which will respect
the configured quota, the server now pushes back requests
if it can't handle them within quota.

Upon receiving a HTTP 429 status code the client should perform
a retry with backoff.

Changes
-------
Add a QueryRetryBehavior interface which can be set at bucket
level to control how queries are retried. Add an implementation
for retry behavior with a delay.

Slightly refactor executeSearchQuery into two functions so that
the HTTP request can be retried.

Change-Id: I34324e36aaeca446ca067b13c26f36f713c33b1c
Reviewed-on: http://review.couchbase.org/99378
Reviewed-by: Ellis Breen <ellis.breen@couchbase.com>
Reviewed-by: Michael Nitschinger <michael@nitschinger.at>
Tested-by: Charles Dixon <chvckd@gmail.com>