GOCBC-350: When credentials are empty don't set basic auth for bucket manager.
Motivation
----------
When an empty string is used for username and password to create a bucket
manager we try to use basic authentication for capiRequest, this returns
a 401. Instead, if username and password are empty then don't set
basic authentication.

Changes
-------
Update capiRequest to only include basic authentication if username
or password are not empty strings.

Change-Id: I0632d8028c4dbb88ffe3f7e89a6b1dc5e2106b0b
Reviewed-on: http://review.couchbase.org/99682
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>