GOCBC-375: Enable analytics and search index tests
Motivation
----------
Now that the changes have gone in for testing against a real server
and feature testing we should uncomment and rewrite, where necessary,
the tests for fts index and for analytics.

Changes
-------
Uncomment and update analytics and search index tests. Do some
refactoring of search index code to be more unit testable.

Change-Id: I9c888eb605dbd5f14d4ecfa9e61ce930743ff8cc
Reviewed-on: http://review.couchbase.org/101696
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>