GOCBC-424: Change ErrNoResults sentinel to a type
Motivation
----------
Currently query services return ErrNoResults for the call to One if
there are no results. We should continue to return an error but one
more in keeping with the error types that we are now adopting.

Changes
-------
Removed ErrNoResults and added a new NoResultsError interface and
implementation, also exposing IsNoResultsError.

Change-Id: I40e479caaaa4dea8412f04ae399f9484e0a1bf75
Reviewed-on: http://review.couchbase.org/107579
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>