GOCBC-427: Change context to cover the whole stream for services
Motivation
----------
At present context is only used for timing out the initial
connection for a service http request. Instead the context should
last the lifetime of the stream.

Changes
-------
Change context creation so that if the user sets a context and
its deadline is before whatever timeout they set then we don’t
touch it and allow it to hard close the connection. If they don’t
(or timeout is shorter) then the context that we create we add 1s to.
Also detect timeout errors on Close.

Change-Id: Iaee0c397e3eed18306c42f48488396ad3be46b65
Reviewed-on: http://review.couchbase.org/107742
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>