GOCBC-430 - Update how timeouts/contexts are selected for KV
Motivation
----------
At present we select the lower of context, operation level timeout,
cluster level timeout. We shouldn't do this, if set then we should
use operation level timeout or context. If both are set then use
the shorter of the two.

Changes
-------
Updated how KV context/timeouts are selected to match the above.

Change-Id: If207e2612ffff7757a15f5d26af6fa6e83affa57
Reviewed-on: http://review.couchbase.org/107899
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>