GOCBC-466: Fix enhanced durability timeouts
Motivation
----------
In calculating enhanced durability timeouts we send a shorter
timeout to the server. We also coerce the timeout to a floor
if the timeout is lower. We should only update the client side
timeout if the value has been coerced.

Changes
-------
Update how we update the context deadline after calculating the
enhanced durability timeout value.

Change-Id: I16a389c54e4b0e79e7f7ebc82f74651ab6088b72
Reviewed-on: http://review.couchbase.org/109250
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>