GOCBC-358: Add  deserializer to query options
Motivation
----------
Sometimes users want to custom transcode query results. We should
support this but as the query data is always returned as JSON we
should make this simpler for the user and just allow them to set
a serializer.

Changes
-------
Add serializer option to query options and use it to deserialize
query results.

Change-Id: Ic12aa58bb15077c78741c36b519e083d1069a0a3
Reviewed-on: http://review.couchbase.org/110279
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>