GOCBC-478: Add  deserializer to analytics options
Motivation
----------
Sometimes users want to custom transcode query results. We should
support this but as the query data is always returned as JSON we
should make this simpler for the user and just allow them to set
a serializer.

Changes
-------
Add serializer option to analytics query options and use it to deserialize
analytics query results.

Change-Id: Ic9fc222e47386606141e271e142f02efcb72ba50
Reviewed-on: http://review.couchbase.org/110282
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>