GOCBC-483: Update transcoder API
Motivation
----------
We should use composition for building the default transcoder with
a serializer. We also shouldn't expose the serializer from the
transcoder.

Changes
-------
Update the signature of NewDefaultTranscoder to take serializer
as a parameter. Update subdoc and services options to
take a serializer directly. Add a serializer to cluster and
options.

Change-Id: I518b41fc9b0c9be1709d95d39be276702a633507
Reviewed-on: http://review.couchbase.org/110355
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>