GOCBC-486: Rename and default analytics options ContextID
Motivation
----------
ContextID on analytics options should be ClientContextID to be
consistent with N1QL and with other documentation. It should also
default to a uuid.

Changes
-------
Changed the name and made the property default to a uuid when converting
to a map at query time. Also moved where the n1ql ClientContextID
is defaulted to a uuid into the toMap function.

Change-Id: I3213c09425ee7705c9d51b7e9074db45aaf3ca7d
Reviewed-on: http://review.couchbase.org/110573
Reviewed-by: Mike Goldsmith <goldsmith.mike@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>