GOCBC-489: Fix how FTS errors are handled
Motivation
----------
We read the errors property on the fts json response from
the top level. This is incorrect as errors lives on the status
object.

Changes
-------
Moved errors within the response struct so that it is part of Status.
Also changed Errors to be an interface so that we can correctly handle
the different ways that FTS throws errors at us. We still maintain
the same public API, only implementation detail has changed.

Change-Id: I50df4d916045bdc2121399323ba6ee012488e894
Reviewed-on: http://review.couchbase.org/110774
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>