GOCBC-514: Don't cache clients that fail to connect
Motivation
----------
We always cache clients in the cluster connections map. This means
that if the user opens multiple buckets then it's possible to end
up with a mix of cached clients that connected and didn't connect.
This means that calls into http services can return a client that
did not connect. It also means that if the user tries to reopen
the bucket then it will fail.

Changes
-------
Change where client connect is called to be performed by cluster.
Only cache the client within the connections if the connection
is successful..

Change-Id: I1a0675af3e08c04afcd2e4c66db1afa7f22e8dfd
Reviewed-on: http://review.couchbase.org/112974
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>