GOCBC-531: Implement query fast prepare support
Motivation
---------
The server now supports "fast prepare" for prepared statements. This
reduces sending uncached prepared statements to the server to a single
round trip.

Changes
-------
Update prepared queries (when enhanced supported) to read the plan
name from the server response and to only send name when the
query is cached.

Change-Id: I5ae1235b7c6c3d0ca291d2a6c08162f78312c30c
Reviewed-on: http://review.couchbase.org/113562
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>