GOCBC-539: Add user roles on Get if there is no origin set
Motivation
----------
If a role does not have an origins field then it means that it's
from an older version of the server and we should take it to
mean that the role belongs to the user not a group.

Changes
-------
Update how users are transformed on Get so that if origin is missing
from a role then we add that role directly to the user.

Change-Id: I8cfd92a0cc749a2fa06b47eb280eb3193f97bdb7
Reviewed-on: http://review.couchbase.org/113618
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>