GOCBC-595: Improve bucket manager error messages
Motivation
----------
When we receive a HTTP response with a non 200 status code sometimes
the response body doesn't contain any information so it's useful if the
error contains the status code. Even when an error message is present it's
useful to also have the status code

Changes
--------
Update the bucket manager invalid status code error handling to always
include the status code. If there is no message in the response
body then omit the Message part of the error and only add the
status code.

Change-Id: I490787f59c23e5632424b5a5398bad17d60cb840
Reviewed-on: http://review.couchbase.org/114338
Reviewed-by: Charles Dixon <chvckd@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>