GOCBC-485: Add support for cluster level operations
Motivation
-----------
As of 6.5 the server supports using connections that haven't had select
bucket run on them. This means that things like queries can be run
without needing to select bucket first.

Changes
-------
Attempt to connect on initial cluster open. If it succeeds then attempt
to use the connection for cluster level operations (fast failing if
the connection does not support GCCCP). When a bucket is opened
then assign the cluster level client to it if it exists and not
already taken by another bucket. On bucket close assign the connection
back to the cluster if it is the final bucket.

Change-Id: Ibe8b001650ed6a1d43aa1f61d0bc065410f1a2c9
Reviewed-on: http://review.couchbase.org/114336
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>