GOCBC-560: Rename subdoc *ops to *spec and make ops top level
Motivation
----------
In order to be more consistent with other SDKs things like
LookupInOp should be LookupInSpec. In order to improve
user experience we should drop the old LookupInSpec and
expose the operations at root level.

Changes
-------
Rename LookupInOp to LookupInSpec. Rename MutateInOp to MutateInSpec.
Remove old LookupInSpec and old MutateInSpec. Move the subdoc op
functions to the root of gocb and rename options blocks to match.

Change-Id: I60e9dccab9965ac49e6c632a973aa6216254c4f1
Reviewed-on: http://review.couchbase.org/115166
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>