GOCBC-616: Fix bug in Get projections with expiry
Motivation
----------
When WithExpiry is used with Get projections and full doc get
then it can lead to a panic because len(ops) != len(result) due
to us removing the first element from result.

Changes
-------
If WithExpiry is set then remove the first op as well as the first
result (they're both the expiration op).

Change-Id: I882df653c2b504c4c8cf3410aaf74a5ae6ffd1f0
Reviewed-on: http://review.couchbase.org/115265
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>