GOCBC-609: Make whole doc projections only project selected fields
Motivation
--------
When the user performs a projection and selects more than 16 fields
then we convert the projection to do a full doc lookup. When this
happens the result should only contain the fields that the user
selected (consider that they might be doing a user fetch and
omitting the password field).

Changes
-------
Update projections so that on full doc fetch (due to too many fields)
the result will only contain the requested fields.

Change-Id: Id2bcfc88a1efea1b1d50460e9e8f9b8a75e568ad
Reviewed-on: http://review.couchbase.org/115322
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>