GOCBC-628: Update ScanConsistency on service options blocks
Motivation
----------
For better consistency and alignment with RFCs all services should
expose ScanConsistency.

Changes
-------
Renamed Consistency to ScanConsistency on Query and Search. Added
the option to Ananlytics. Also renamed the Query ConsistencyMode
to QueryScanConsistency and created corresponding types for Search
and Analytics. Tidied up related error messages to be
configurationError.

Change-Id: Ibe5c69c853c3a3aa9d6c717433d07b65cbb01927
Reviewed-on: http://review.couchbase.org/115651
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>