GOCBC-656: Add Threshold Logging and tracing to KV
Motivation
----------
As a part of adding RTO to the SDK we need to add threshold logging.

Changes
-------
Added an internal tracing API and created a threshold logger
and a noop logger that implement it. Default to threshold logger
if not disabled. Wire tracing into most KV ops without exposing
publicly for now.

Change-Id: I0534be19be5f50fa2a693b487bd65e8b1aed6c5c
Reviewed-on: http://review.couchbase.org/117026
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>