GOCBC-676: Updated Exists to use GetMeta instead of Observe.
Motivation
----------
Server changes in 6.5 have made it such that Observe may not
always return the appropriate answer if it is used to test the
existence of a document.  To work around this, SDKs now use
GetMeta to fetch this information.

Changes
-------
Updated Exists() to use the GetMeta operation instead of
Observe.  Updated the mockKvProvider for this operation
and disabled a few tests that used Exists when running on
the mock as it does not support GetMeta.

Change-Id: I9afec5b194c991ac4919f3c44edf1b6325983828
Reviewed-on: http://review.couchbase.org/120394
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Brett Lawson <brett19@gmail.com>