GOCBC-787: Fix subdoc ExpandMacros and blank path Replace
Motivation
----------
Expand macros tests are failing as are tests using Replace with a
blank path. The server is rejecting requests for either test.

Changes
-------
Ensure that XAttr flag is set if ExpandMacros is. Ensure that the
macros are sent in quotes. If path is blank for a Replace then
ensure that we send the correct op.

Change-Id: I4d6d350cbcce4174f29e49554d944c968f37b974
Reviewed-on: http://review.couchbase.org/122043
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>