GOCBC-785: Add a new node version and update tests feature support
Motivation
----------
Just because a server version is 6.5 does not mean that developer
preview has been enabled and collections are supported. We should
add a new supported cluster version and only run collections
tests if that is used.

Changes
-------
Only run collections tests if the node version specified (note that is
not necessarily the actual cluster version) is 7.0 or up.

Change-Id: I7d8b833ec29051d4fb44b20827d7fbb44c327763
Reviewed-on: http://review.couchbase.org/122044
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>