GOCBC-722: Use testify mock for mocking
Motivation
----------
Our mocks are now out of date in some places and they are hard to
use. As we implemented more unit tests we are going to need better
mocking. We should use testify mocks for this.

Changes
-------
Removed all of our own mock implementations and replaced with
testify mocks, wiring them up appropriately.

Change-Id: I0ae19a74a6ebffd221fd09f193a0a6e28d174cc0
Reviewed-on: http://review.couchbase.org/122124
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>