GOCBC-788: Update TestTouch test to only test Touch
Motivation
----------
The TestTouch test is testing both GetAndTouch and Touch, the
logic between the sleeps and exiries set is also incorrect
which is leading the test to fail.

Changes
-------
Update TestTouch to not test GetAndTouch and remove all sleeps
preferring to use a Get WithExpiry to confirm that an expiry
has been set.

Change-Id: I6286749728e9246373bf1d1370d539e7f9d9b268
Reviewed-on: http://review.couchbase.org/122161
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>