GOCBC-794: Use correct addresses in cluster manager
Motivation
----------
When connecting to a cluster using a non default port and the
couchbase:// scheme we can trigger a panic when trying to use
`cluster.Manager(username,password).Internal().GetNodesMetadata()`.
This is triggering because we aren't pulling the http addresses
from the agent and instead pull them from the supplied
agentconfig (so no http addresses...).
We also have no guards in place to prevent a nil pointer if there
are no http addresses to use.

Changes
-------
Get the http addresses from the agent (which will ensure that they
are always current) and add guards to prevent a nil pointer occurring
if there are no addresses to use.

Change-Id: I630813762e27ea5f0fe1d4d1d738164d9e6fd456
Reviewed-on: http://review.couchbase.org/122334
Tested-by: Ben Brooks <ben.brooks@couchbase.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>