GOCBC-716 - Update bucket manager error handling
Motivation
----------
When flush is called on bucket and flush is disabled we don't
return ErrBucketNotFlushable and we should. We should
also improve the rest of the error handling to return typed errors
where possible and meaningful errors where not.

Changes
--------
Update bucket manager to parse response bodies and try to extract
and error information.

Change-Id: I88605c095d3a3f73037188a917bf126f5ca5c838
Reviewed-on: http://review.couchbase.org/122495
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>