GOCBC-808: Retry 4040 code on prepared statement execution
Motivation
----------
Due to scenarios during an upgrade to server version 6.5 it's
possible for the server to return a 4040 to us when we send a
prepared statement. When this happens we should reprepare the
statement.

Changes
-------
Add 4040 to the list of error codes that we check after sending
a prepared statement name/plan. If a 4040 is found then reprepare.
happens we should

Change-Id: I5c617032bd8b8edf030afde62208fed2ea2c16af
Reviewed-on: http://review.couchbase.org/122705
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>