GOCBC-721: Update manager implementations to be consistent
Motivation
----------
The SDK contains a number of managers. Each of them is implemented
differently, some of them take a provider and various configuration
options, which are then used to allow the manager to work
independently of its source object, while some managers take
references to the source object and work through them
instead.
All the managers should be consistent in their implementation.

Changes
-------
Update managers so that they all rely on interfaced providers
rather than cluster or bucket objects directly.

Change-Id: Ifbaa07f7ba597f2ec6f84d7ee6fa7cae76d84e78
Reviewed-on: http://review.couchbase.org/122720
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>