GOCBC-817: Don't reconnect clients if they already exist
Motivation
----------
When a Bucket call is used on the same named bucket twice then the
underlying connection created on the first call should be shared across
both buckets. At present we do not correctly set the bucket name on
the client on the first bucket call and then on the second call we
correctly get the existing connection but we then try to reconnect it
and we do the reconnection with no bucket name set in the config.
This leads to operations on both buckets returning errors that no
bucket has been selected.

Changes
-------
Correctly set the bucket name on the client when select bucket
is performed. If there is already a client connection for a given
bucket name then use that connection but don't try to reconnect it.

Change-Id: I42a63ff34129cad37e7d7c91bc55c1be746bb5f6
Reviewed-on: http://review.couchbase.org/123127
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>