GOCBC-821: Fallback to using supplied http addresses in cluster manager
Motivation
----------
We recently made a fix which meant that we use the correct http addresses
in the cluster manager by pulling them off of the agent. This has
broken existing behaviour that some users were using (management of
clusters to add buckets from fresh as one example). We should add
a fallback that if no agents exist then we'll try to use
any http addresses specified.

Changes
-------
In cluster manager requests try to pull http addresses off of the
agent. If no agent exists then fallback to trying to use the http
addresses parsed from the connection string.

Change-Id: I60ebf3217de349b4024bb68da32bb7a7d5c5674d
Reviewed-on: http://review.couchbase.org/123158
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>