GOCBC-820: Remove unused context.Context option from search indexes
Motivation
----------
We currently have context.Context in search index manager options
blocks. This is a hangover from earlier during the alpha stage of
the SDK and is totally unused from within the SDK.
This is technically a breaking change however it's better to
break it in order to ensure that any users using it are made
aware that it is not doing what they may think.

Changes
-------
Remove context.Context from search index manager options blocks.

Change-Id: I974cbb9017b25f29c98f546163652d04fc150720
Reviewed-on: http://review.couchbase.org/123234
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>