Update GetAllReplicas test
Motivation
----------
The GetAllReplicas test randomly fails on Jenkins. This is because
we set PersistTo to be hardcoded to 1, and we then check that
we've had a result from all servers. We should set persistTo to
the number of servers available.

Changes
-------
Update GetAllReplicas test to persistTo the number of servers in
the cluster rather than hardcoded to 1.

Change-Id: I74f5a8bbe17df621602c91fb4215b1f4ed3e9c8b
Reviewed-on: http://review.couchbase.org/123433
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>