GOCBC-839: Update UpdateBucket to use correct uri
Motivation
----------
At the moment UpdateBucket is basically an alias of InsertBucket,
this means that it uses the same uri. This is incorrect as update
should append the bucket name to the end of the uri.

Changes
-------
Created a private function that Insert and Update both call but
uses the correct uri depending on the behaviour required.

Change-Id: I77394b1b62548cb19d13181e9232f797c6e2d1a8
Reviewed-on: http://review.couchbase.org/c/gocb/+/124705
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>