Add support for DP edition and update test support
Motivation
----------
At present the 6.5 DP mode has an issue with indexing which causes
our tests to fail against it. We should add a DP edition to our
node versions and not run query, analytics etc... tests against
it.

Changes
-------
Add a modifier flag that can be set to DP and don't run query
etc... tests if it is in use.
Change-Id: Iec6e940409843cf13e16d3f64e3268f94c9266c1
Reviewed-on: http://review.couchbase.org/c/gocb/+/125350
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>