Only use connections that are valid for HTTP based services
Motivation
----------
Now that we do not receive connect time errors from gocbcore we
need to alter how we select connections to use when performing
any HTTP based requests.

Changes
-------
Update how we select connections so that only ones that are connected
are used.
Update the kvProvider interface to match gocbcore.

Change-Id: I91069818065dd03861906bd79c5cb395838fd69c
Reviewed-on: http://review.couchbase.org/c/gocb/+/126173
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>