GOCBC-879: Use gocbcore AgentGroup
Motivation
----------
We try to manage multiple agents from within gocb and when we perform
http requests (or n1ql etc...) we attempt to work out which, if any,
agents are connected. This logic is fragile and often results in
misleading errors being returned to users if WaitUntilReady is not
used. Gocbcore now has the concept of AgentGroup which will handle
multiple agents for us and will also block http based requests until
connections are ready, the same as KV.

Changes
-------
Update client and providers to use AgentGroup.
Update cluster to not try to work out if clients are connected or
if they're using GCCCP.
Rename client interface to connectionManager to make the intent
more clear.

Change-Id: Icefaaf81cc434f76f1ae64e2e663b745d8d82e05
Reviewed-on: http://review.couchbase.org/c/gocb/+/129183
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>