GOCBC-900: Respect QueryIndex Ignore options
Motivation
----------
We provide options to ignore if an index already exists when creating
an index. We do the same for if the index doesn't exist when dropping.
We don't actually do anything with these options.

Changes
-------
Update logic so that if an index already exists and ignore is set
then don't return an error.
Same as above for if the index doesn't exist on drop.

Change-Id: I3f5322ebd5d5dd94d2e22c36728c9c3a1a591c83
Reviewed-on: http://review.couchbase.org/c/gocb/+/129528
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>