GOCBC-902: Improve linting abilities
Motivation
----------
We've been using the same linting for quite some time now and IDEs
are beginning to flag up issues that our CI isn't. We should improve
our linting.

Changes
-------
Replace all existing linting with golangci-lint which covers the
linters we were using and many more.
Pick a sensible set of linters for now, allowing us to add more in
the future.
Make various code fixes to satisfy the linters.

Change-Id: Ic74071f24975c9ba756713051647e6355c779ae6
Reviewed-on: http://review.couchbase.org/c/gocb/+/129730
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>