GOCBC-913: If no Ping services specified don't error for missing services
Motivation
-----------
At the moment if a user does a Ping specifying no services against
a cluster which does not have a service such as analytics then we
will include that service in the results as unavailable. We should
instead only ping services that are in the cluster config if no
services are specified.

Changes
-------
If no service types set on PingOptions then don't set any before
sending to gocbcore, let gocbcore decide which to perform.
Update gocbcore version in gomod.

Change-Id: Idd31b452a5d69c9687f0f357139dab549b96eb81
Reviewed-on: http://review.couchbase.org/c/gocb/+/130178
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>