GOCBC-718: Expose search/view parsing errors through .Err
Motivation
----------
At the moment if there's an error parsing a search or view row then
we just swallow it. This is something that should never happen but
we should expose it in some way.

Changes
--------
If there's a parsing error then store it and make it accessible via
the .Err call.

Change-Id: Ie1200ed22e28dc9827508fdde8d4d051688aac96
Reviewed-on: http://review.couchbase.org/c/gocb/+/131561
Tested-by: Charles Dixon <chvckd@gmail.com>
Reviewed-by: Brett Lawson <brett19@gmail.com>