GOCBC-962: Expose ThresholdLoggingTracer
Motivation
----------
When creating SDK 2.0 we moved the threshold logging tracer over
from 1.x but we did not expose it via the public API. It is
still the default tracer that we use if the user does not provide a
tracer implementation but we expose no way to configure it.

Changes
-------
Expose ThresholdLoggingTracer so that users can configure it and
set it using the Tracer field in ClusterOptions.

Change-Id: I4bca29c3aafbb0919b1a594ba902cd22dd2eb82d
Reviewed-on: http://review.couchbase.org/c/gocb/+/133380
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>