GOCBC-970: Run query, search, view tests against default collection
Motivation
----------
The indexes for query, analytics etc... aren't indexing documents
created outside of the non-default collection because we create
indexes on the default collection. We should create documents
on the default collection and as we add support for collections
with the services we should add tests for non-default collection.

Changes
-------
Update tests to run query, analytics, search, views against the
default collection.

Change-Id: I37a1b95b643ffc8df11aa3a2d7e8141165a65cce
Reviewed-on: http://review.couchbase.org/c/gocb/+/134432
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>