GOCBC-969: Only try to parse query metrics if present
Motivation
----------
We always attempt to parse the query metrics, even if they aren't
present. This doesn't result in an error but does log out a message.
We should just not try to parse metrics if they aren't in the body.

Changes
-------
Only try to parse query metrics if they're present in the result
body.

Change-Id: Ie04673825093641c383c0a0f14d429d8a815db04
Reviewed-on: http://review.couchbase.org/c/gocb/+/134718
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Charles Dixon <chvckd@gmail.com>